+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Release 4
 =========
 
-* Rename the scriptes:
-  * c2cwsgiutils_run =>
+* Rename the scripts:
+  * c2cwsgiutils_run => c2cwsgiutils-run
   * c2cwsgiutils_genversion.py => c2cwsgiutils-genversion
   * c2cwsgiutils_coverage_report.py => c2cwsgiutils-coverage-report
   * c2cwsgiutils_stats_db.py => c2cwsgiutils-stats-db